requirements.txt
blobfile
imageio
mpi4py
numpy
pillow
tensorboard
torch
tqdm
